Refactor fuel price functions: move fetch_fuel_prices and modify_fuel_prices to fuels.py, update imports in generators.py and fuel_test.py
